% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPtransfo.R
\name{barW}
\alias{barW}
\title{Population mean given MC samples of latent variables}
\usage{
barW(MC)
}
\arguments{
\item{MC}{MC samples of l1 and l2}
}
\value{
None
}
\description{
population mean given MC samples of latent variables
}
\examples{

}
