% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPtransfo.R
\name{Ey}
\alias{Ey}
\title{Expectation of a zero-inflated Poisson model}
\usage{
Ey(l1, l2)
}
\arguments{
\item{l1}{Latent linear predictor value for the conditional Poisson part of a model}

\item{l2}{Latent linear predictor value for the probability of a true zero}
}
\value{
None
}
\description{
Expectation of a zero-inflated Poisson model from its two latent linear predictors.
}
\examples{

}
