% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListHandle.R
\name{meltlists}
\alias{meltlists}
\title{Combine two lists}
\usage{
meltlists(x, y)
}
\arguments{
\item{x}{A list of vector and/or matrices}

\item{y}{Another list of same dimension as x, with elements matching those of x}
}
\value{
None
}
\description{
Combine two lists of vectors or matrices
}
\examples{
x <- list(matrix(0,10,10), matrix(1,10,10), rnorm(100))
y <- list(matrix(2,10,10), matrix(3,10,10), rnorm(100))
meltlists(x=x, y=y)

}
