% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPtransfo.R
\name{Va_w_zip}
\alias{Va_w_zip}
\title{Additive genetic variance in relative fitness from ZIP model sample}
\usage{
Va_w_zip(m, s = 1, nMC = 50000, h = 0.02, fitnessfile = NULL,
  fixedpred = c("inbreeding", "Qgg"), ranpred = c("id", "dam",
  "cohort"), focalvar = "id")
}
\arguments{
\item{m}{MCMCglmm model}

\item{s}{MCMC sample}

\item{nMC}{number of MCMC sample to use for integration}

\item{h}{Deviations to use for derivation}

\item{fitnessfile}{CSV file containing fitness data (in a column called LBS)}

\item{fixedpred}{Fixed effects}

\item{ranpred}{Random effects}

\item{focalvar}{Name of the random effects to back-transform}
}
\value{
None
}
\description{
Function to give the genetic variance of relative fitness given ZIP model M for MCMC sample s.  This is itself done by MC integration, with nMC samples
}
\examples{

}
