% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPtransfo.R
\name{Va_w_zip_posterior}
\alias{Va_w_zip_posterior}
\title{Back-transform variances in relative fitness from ZIP model}
\usage{
Va_w_zip_posterior(m, nMC = 50000, h = 0.02, fitnessfile,
  fixedpred = c("inbreeding", "Qgg"), ranpred = c("id", "dam",
  "cohort"))
}
\arguments{
\item{m}{MCMCglmm model}

\item{nMC}{number of MCMC sample to use for integration}

\item{h}{Deviations to use for derivation}

\item{fitnessfile}{CSV file containing fitness data (in a column called LBS)}

\item{fixedpred}{Fixed effects}

\item{ranpred}{Random effects}
}
\value{
None
}
\description{
Function to give the genetic variance of relative fitness given ZIP model.
}
\examples{

}
