\name{VAiswow-package}
\alias{VAiswow-package}
\alias{VAiswow}
\docType{package}
\title{
\packageTitle{VAiswow}
}
\description{
\packageDescription{VAiswow}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VAiswow}
\packageIndices{VAiswow}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VAiswow}

Maintainer: \packageMaintainer{VAiswow}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
