% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPtransfo.R
\name{zip_pred}
\alias{zip_pred}
\title{Model prediction from a MCMC sample of a ZIP model}
\usage{
zip_pred(m, s = 1, nMC = 50, fitnessfile = NULL,
  fixedpred = c("inbreeding", "Qgg"), ranpred = c("id", "dam",
  "cohort"))
}
\arguments{
\item{m}{MCMCglmm model}

\item{s}{MCMC sample}

\item{nMC}{number of MCMC sample to use for integration}

\item{fitnessfile}{CSV file containing fitness data (in a column called LBS)}

\item{fixedpred}{Fixed effects}

\item{ranpred}{Random effects}
}
\value{
None
}
\description{
Function to give the expected distribution of LBS given a ZIP model and fixed effect predictor values
}
\examples{

}
